style: rename useCurrentJobRun to useActiveJobRun
